Add configure_nmake macro
The macro utilises bazel "transitions" to set the `make` toolchain to
be the preinstalled `nmake` tool.

Note that the msvc constraint was removed from the
`exec_compatible_with` attribute of `preinstalled_nmake_toolchain` as
the condition is not actually met even when building with msvc. See
https://github.com/bazelbuild/bazel/issues/7730.

This will be tested in PR#729